Move common method for a CapsuleManager to trait, this would reduce the requirement to produce new CapsuleManager for other component in the future (e.g: laravel/framework#5032)
Also remove duplicate "require-dev" component (when already explicitly declare in "require").
