+++ src/Illuminate/Database/Capsule/Manager.php
@@ -1,22 +1,17 @@
 <?php namespace Illuminate\Database\Capsule;
 
 use PDO;
-use Illuminate\Support\Fluent;
 use Illuminate\Events\Dispatcher;
 use Illuminate\Cache\CacheManager;
 use Illuminate\Container\Container;
 use Illuminate\Database\DatabaseManager;
 use Illuminate\Database\Eloquent\Model as Eloquent;
 use Illuminate\Database\Connectors\ConnectionFactory;
+use Illuminate\Support\Traits\CapsuleManagerTrait;
 
 class Manager {
 
-	/**
-	 * The current globally used instance.
-	 *
-	 * @var \Illuminate\Database\Capsule\Manager
-	 */
-	protected static $instance;
+	use CapsuleManagerTrait;
 
 	/**
 	 * The database manager instance.
@@ -24,13 +19,6 @@ class Manager {
 	 * @var \Illuminate\Database\DatabaseManager
 	 */
 	protected $manager;
-
-	/**
-	 * The container instance.
-	 *
-	 * @var \Illuminate\Container\Container
-	 */
-	protected $container;
 
 	/**
 	 * Create a new database capsule manager.
@@ -48,22 +36,6 @@ class Manager {
 		$this->setupDefaultConfiguration();
 
 		$this->setupManager();
-	}
-
-	/**
-	 * Setup the IoC container instance.
-	 *
-	 * @param  \Illuminate\Container\Container|null  $container
-	 * @return void
-	 */
-	protected function setupContainer($container)
-	{
-		$this->container = $container ?: new Container;
-
-		if ( ! $this->container->bound('config'))
-		{
-			$this->container->instance('config', new Fluent);
-		}
 	}
 
 	/**
@@ -183,16 +155,6 @@ class Manager {
 	}
 
 	/**
-	 * Make this capsule instance available globally.
-	 *
-	 * @return void
-	 */
-	public function setAsGlobal()
-	{
-		static::$instance = $this;
-	}
-
-	/**
 	 * Get the database manager instance.
 	 *
 	 * @return \Illuminate\Database\Manager
@@ -248,27 +210,6 @@ class Manager {
 	public function setCacheManager(CacheManager $cache)
 	{
 		$this->container->instance('cache', $cache);
-	}
-
-	/**
-	 * Get the IoC container instance.
-	 *
-	 * @return \Illuminate\Container\Container
-	 */
-	public function getContainer()
-	{
-		return $this->container;
-	}
-
-	/**
-	 * Set the IoC container instance.
-	 *
-	 * @param  \Illuminate\Container\Container  $container
-	 * @return void
-	 */
-	public function setContainer(Container $container)
-	{
-		$this->container = $container;
 	}
 
 	/**
